Implementação das ondulações

+++ leitura_entrada.h
@@ -1,1 +1,1 @@
-void leia_entrada(char*, int*, int*, int*, int*, int*, int*, int*, int*, float*, float*);
+void leia_entrada(char*, int*, int*, int*, int*, int*, int*, int*, int*, double*, double*);
